create security question base class, extend to reuse on registration flow
